refactor: Use CSS Modules to style CopyableHeader.
