+++ components/blog/default-replacements.tsx
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import { CopyableHeader } from '../CopyableHeader';
+import { CopyableHeader } from '../CopyableHeader/CopyableHeader';
 import { Link } from '../Link/Link';
 
 // https://github.com/hashicorp/next-mdx-remote#replacing-default-components
